Add initial assembly_scanner
assembly_scanner and program_walker should inherit from a common generic
abstract class, and obviously a *ton* of testing is needed for
assembly_scanner itelf. Once that's in place, thread can use a generic
form of IP and advancement, and threads can be held by a subclass of
either of the above.